combotsha.py: _Repository.get_new_commits(): return oldest commit first
This patch makes _Repository.get_new_commits() reverse the returned list
so that the first item is the oldest commit.

This effectively sorts the IRC output so that, if you push a batch of
four commits, then the bot prints:

    barectf: 2020-09-03 21:41: [Philippe Proulx] c268d669 cgen.py: _CodeGen.gen_src(): remove ugly empty lines before `}` (+27 -6)
    barectf: 2020-09-03 21:45: [Philippe Proulx] 9aeb842c barectf.c.j2: remove superfluous empty line (+0 -1)
    barectf: 2020-09-03 21:53: [Philippe Proulx] dfb4579c config_parse_v3.py: fix Pylint errors/warnings (+0 -1)
    barectf: 2020-09-03 22:03: [Philippe Proulx] 18838ea3 barectf: fix Flake8 errors (+0 -1)

This order is also what you get when you do perform an interactive
rebase.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I8433e13cd770d20c597523363a9f63b34fa4bac8